don't regen localhost certs for every test

+++ testing/src/exec/certs.rs
@@ -1,3 +1,5 @@
+use std::sync::Mutex;
+use std::time::{Duration, SystemTime};
 use std::{fs, io, path::PathBuf, process::Command};
 
 #[derive(Debug)]
@@ -7,8 +9,33 @@ pub struct Certificates {
     pub cert_file_der: PathBuf,
 }
 
+static CERT_CREATION: Mutex<()> = Mutex::new(());
+
 /// Creates a self signed cert & key for localhost in the specified directory.
-pub fn create_localhost_key_and_cert(dir: PathBuf) -> io::Result<Certificates> {
+pub fn create_localhost_key_and_cert(dir: PathBuf, force: bool) -> io::Result<Certificates> {
+    let certs = Certificates {
+        key_file_pem: dir.join("localhost.key"),
+        cert_file_pem: dir.join("localhost.cert"),
+        cert_file_der: dir.join("localhost.cert.der"),
+    };
+
+    let _locked = CERT_CREATION.lock().unwrap();
+    let file_ok = |f: &PathBuf| -> io::Result<bool> {
+        Ok(f.exists()
+            && f.is_file()
+            && SystemTime::now()
+                .duration_since(f.metadata()?.created()?)
+                .unwrap()
+                < Duration::from_secs(60 * 60 * 24 * 300))
+    };
+    if !force
+        && file_ok(&certs.key_file_pem).is_ok_and(|v| v)
+        && file_ok(&certs.cert_file_pem).is_ok_and(|v| v)
+        && file_ok(&certs.cert_file_der).is_ok_and(|v| v)
+    {
+        return Ok(certs);
+    }
+
     let cfg = include_str!("openssl_req.txt");
     let config_file = dir.join("openssl_req.txt");
     fs::write(config_file, cfg.as_bytes())?;
@@ -53,9 +80,5 @@ pub fn create_localhost_key_and_cert(dir: PathBuf) -> io::Result<Certificates> {
         .status()
         .expect("couldn't covert cert PEM to DER");
 
-    Ok(Certificates {
-        key_file_pem: dir.join("localhost.key"),
-        cert_file_pem: dir.join("localhost.cert"),
-        cert_file_der: dir.join("localhost.cert.der"),
-    })
+    Ok(certs)
 }
